# docs/API_SPEC/openapi.yaml
openapi: 3.1.0
info:
  title: FinPulse API
  description: |
    REST API for the FinPulse real-time wealth management and high-frequency trade auditing platform.

    ## Authentication
    All endpoints (except `/api/v1/auth/**`) require a Bearer JWT in the `Authorization` header.
    Tokens are obtained via `POST /api/v1/auth/login` and expire after **15 minutes**.
    Use `POST /api/v1/auth/refresh` with the `refreshToken` HttpOnly cookie to obtain a new access token.

    ## Rate Limiting
    1000 requests per minute per IP address. Exceeding this returns HTTP 429.

    ## Error Format
    All errors follow [RFC 9457 Problem Details](https://www.rfc-editor.org/rfc/rfc9457):
    ```json
    { "type": "...", "title": "...", "status": 400, "detail": "...", "traceId": "..." }
    ```
  version: 1.0.0
  contact:
    name: FinPulse Team
    email: dev@finpulse.dev

servers:
  - url: https://api.finpulse.dev
    description: Production
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Authentication and session management
  - name: Wallets
    description: Wallet balance and ledger operations
  - name: Trades
    description: Trade order placement and history
  - name: Market
    description: Market data and price history
  - name: Audit
    description: Compliance and audit log access (AUDITOR / ADMIN only)
  - name: Admin
    description: Platform administration (ADMIN only)

# ─────────────────────────────────────────────────────────────────────────────
# SECURITY SCHEMES
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── SHARED SCHEMAS ───────────────────────────────────────────────────────
  schemas:

    ProblemDetail:
      type: object
      properties:
        type:    { type: string, example: "https://finpulse.dev/errors/validation" }
        title:   { type: string, example: "Validation Failed" }
        status:  { type: integer, example: 400 }
        detail:  { type: string, example: "Field 'quantity' must be positive" }
        traceId: { type: string, example: "a1b2c3d4e5f6" }

    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 50, example: "shriansh" }
        email:    { type: string, format: email, example: "shri@finpulse.dev" }
        password: { type: string, minLength: 8, example: "MyStr0ng!Pass" }

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: "shriansh" }
        password: { type: string, example: "MyStr0ng!Pass" }

    AuthResponse:
      type: object
      properties:
        accessToken: { type: string, example: "eyJhbGciOiJSUzI1NiJ9..." }
        expiresIn:   { type: integer, example: 900, description: "Seconds until token expiry" }
        userId:      { type: string, format: uuid }
        roles:       { type: array, items: { type: string }, example: ["ROLE_USER"] }

    WalletDto:
      type: object
      properties:
        id:               { type: string, format: uuid }
        userId:           { type: string, format: uuid }
        currency:         { type: string, example: "USD" }
        balance:          { type: number, format: double, example: 1000.00 }
        lockedBalance:    { type: number, format: double, example: 500.00 }
        availableBalance: { type: number, format: double, example: 500.00 }

    DepositRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:   { type: number, format: double, minimum: 0.01, example: 500.00 }
        currency: { type: string, default: "USD", example: "USD" }

    WithdrawRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:   { type: number, format: double, minimum: 0.01, example: 200.00 }
        currency: { type: string, default: "USD", example: "USD" }

    TradeRequest:
      type: object
      required: [symbol, side, quantity, price]
      properties:
        symbol:   { type: string, pattern: "^[A-Z]{2,10}/[A-Z]{2,10}$", example: "BTC/USD" }
        side:     { type: string, enum: [BUY, SELL] }
        quantity: { type: number, format: double, minimum: 0.00000001, example: 0.05 }
        price:    { type: number, format: double, minimum: 0.01, example: 60000.00 }

    TradeResponse:
      type: object
      properties:
        id:          { type: string, format: uuid }
        userId:      { type: string, format: uuid }
        symbol:      { type: string, example: "BTC/USD" }
        side:        { type: string, enum: [BUY, SELL] }
        quantity:    { type: number, format: double }
        price:       { type: number, format: double }
        totalCost:   { type: number, format: double }
        status:      { type: string, enum: [PENDING, COMPLETED, FAILED, CANCELLED] }
        createdAt:   { type: string, format: date-time }
        executedAt:  { type: string, format: date-time, nullable: true }

    PagedTradesResponse:
      type: object
      properties:
        content:       { type: array, items: { $ref: '#/components/schemas/TradeResponse' } }
        page:          { type: integer, example: 0 }
        size:          { type: integer, example: 20 }
        totalElements: { type: integer, example: 145 }
        totalPages:    { type: integer, example: 8 }

    OhlcvCandle:
      type: object
      properties:
        bucket: { type: string, format: date-time }
        symbol: { type: string, example: "BTC/USD" }
        open:   { type: number, format: double }
        high:   { type: number, format: double }
        low:    { type: number, format: double }
        close:  { type: number, format: double }
        volume: { type: number, format: double }

    AuditLogEntry:
      type: object
      properties:
        timestamp:  { type: string, format: date-time }
        id:         { type: string, format: uuid }
        userId:     { type: string, format: uuid }
        action:     { type: string, example: "TRADE_PLACED" }
        entityType: { type: string, example: "TRADE" }
        entityId:   { type: string, format: uuid }
        details:    { type: object, additionalProperties: true }
        ipAddress:  { type: string, example: "192.168.1.1" }

    PagedAuditResponse:
      type: object
      properties:
        content:       { type: array, items: { $ref: '#/components/schemas/AuditLogEntry' } }
        page:          { type: integer }
        size:          { type: integer }
        totalElements: { type: integer }
        totalPages:    { type: integer }

    PortfolioAsset:
      type: object
      properties:
        symbol:         { type: string, example: "BTC/USD" }
        quantity:       { type: number, format: double }
        avgEntryPrice:  { type: number, format: double }
        currentPrice:   { type: number, format: double }
        marketValue:    { type: number, format: double }
        unrealizedPnl:  { type: number, format: double }
        pnlPercent:     { type: number, format: double }

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:

  # ─── AUTH ───────────────────────────────────────────────────────────────────

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Validation error or email already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and obtain JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: "refreshToken=<uuid>; HttpOnly; Secure; SameSite=Strict; Max-Age=604800"
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }

  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Obtain a new access token using refresh token cookie
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Refresh token expired, revoked, or missing

  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke current JWT
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logged out; token added to Redis blacklist
        '401':
          description: Token missing or already invalid

  # ─── WALLETS ────────────────────────────────────────────────────────────────

  /api/v1/wallets/user/{userId}/{currency}:
    get:
      tags: [Wallets]
      summary: Get wallet balance for a user
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: currency
          in: path
          required: true
          schema: { type: string, example: "USD" }
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletDto' }
        '403':
          description: Access denied (not owner or admin)
        '404':
          description: Wallet not found

  /api/v1/wallets/deposit:
    post:
      tags: [Wallets]
      summary: Deposit funds into wallet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DepositRequest' }
      responses:
        '200':
          description: Updated wallet after deposit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletDto' }
        '400':
          description: Invalid amount

  /api/v1/wallets/withdraw:
    post:
      tags: [Wallets]
      summary: Withdraw funds from wallet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WithdrawRequest' }
      responses:
        '200':
          description: Updated wallet after withdrawal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletDto' }
        '400':
          description: Insufficient available funds

  # ─── TRADES ─────────────────────────────────────────────────────────────────

  /api/v1/trades:
    post:
      tags: [Trades]
      summary: Place a BUY or SELL trade order
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, maxLength: 128 }
          description: Unique key to prevent duplicate order submission
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TradeRequest' }
      responses:
        '202':
          description: Order accepted and queued for processing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeResponse' }
        '400':
          description: Validation error or insufficient funds
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetail' }
        '403':
          description: KYC not approved

  /api/v1/trades/user/{userId}:
    get:
      tags: [Trades]
      summary: Get paginated trade history for a user
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, COMPLETED, FAILED, CANCELLED] }
      responses:
        '200':
          description: Paginated trade list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedTradesResponse' }
        '403':
          description: Access denied

  /api/v1/trades/{tradeId}:
    get:
      tags: [Trades]
      summary: Get a single trade by ID
      security:
        - BearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Trade details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TradeResponse' }
        '403':
          description: Access denied (not owner or admin)
        '404':
          description: Trade not found

  # ─── MARKET ─────────────────────────────────────────────────────────────────

  /api/v1/market/symbols:
    get:
      tags: [Market]
      summary: Get list of supported trading symbols
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of symbol strings
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
                example: ["BTC/USD", "ETH/USD", "AAPL", "TSLA"]

  /api/v1/market/{symbol}/history:
    get:
      tags: [Market]
      summary: Get OHLCV candle history for a symbol
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string, example: "BTC/USD" }
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [1D, 1W, 1M, 1Y]
            default: 1D
      responses:
        '200':
          description: Array of OHLCV 1-minute candles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OhlcvCandle' }
        '404':
          description: Symbol not found

  /api/v1/market/{symbol}/price:
    get:
      tags: [Market]
      summary: Get latest price for a symbol (from Redis cache)
      security:
        - BearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string, example: "BTC/USD" }
      responses:
        '200':
          description: Latest price
          content:
            application/json:
              schema:
                type: object
                properties:
                  symbol:    { type: string }
                  price:     { type: number, format: double }
                  timestamp: { type: string, format: date-time }

  # ─── PORTFOLIO ──────────────────────────────────────────────────────────────

  /api/v1/portfolio/user/{userId}:
    get:
      tags: [Trades]
      summary: Get portfolio asset holdings for a user
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Array of portfolio asset positions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PortfolioAsset' }
        '403':
          description: Access denied

  # ─── AUDIT ──────────────────────────────────────────────────────────────────

  /api/v1/audit/user/{userId}:
    get:
      tags: [Audit]
      summary: Get paginated audit log for a user (AUDITOR / ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date, example: "2026-01-01" }
        - name: to
          in: query
          schema: { type: string, format: date, example: "2026-06-01" }
        - name: action
          in: query
          schema: { type: string, example: "TRADE_PLACED" }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Paginated audit log
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedAuditResponse' }
        '403':
          description: Insufficient role

  /api/v1/audit/export:
    get:
      tags: [Audit]
      summary: Export audit log as CSV (ADMIN only)
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: CSV file stream
          content:
            text/csv:
              schema: { type: string, format: binary }
        '403':
          description: Admin only

  # ─── ADMIN ──────────────────────────────────────────────────────────────────

  /api/v1/admin/users/{userId}/role:
    patch:
      tags: [Admin]
      summary: Update a user's role (ADMIN only)
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [ROLE_USER, ROLE_ANALYST, ROLE_AUDITOR, ROLE_ADMIN]
      responses:
        '200':
          description: Role updated
        '403':
          description: Admin only
        '404':
          description: User not found

  /api/v1/admin/users/{userId}/kyc:
    patch:
      tags: [Admin]
      summary: Update a user's KYC status (ADMIN only)
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kycStatus]
              properties:
                kycStatus:
                  type: string
                  enum: [PENDING, APPROVED, REJECTED]
      responses:
        '200':
          description: KYC status updated
        '403':
          description: Admin only

  # ─── HEALTH ─────────────────────────────────────────────────────────────────

  /actuator/health:
    get:
      tags: [Admin]
      summary: Application health check (public)
      responses:
        '200':
          description: Application is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: "UP" }